pyproject.toml
[project]
name = "quine"
version = "0.1.0"
description = "Literate programming environment — source and prose connected by named tags"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "markdown>=3.5",
    "pygments>=2.17",
]

[project.scripts]
quine = "quine:main"

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "collect", "render"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.4",
]
